Move IntervalSet out of the innards into the public API (#289)
CurrentState::copy_of_values() returns an IntervalSet by value from a
documented public method, every header installs, and gch::small_vector's
include directories already propagate to consumers because this header
is reachable from the public surface — so the type was already de facto
public. Make that deliberate: it becomes gcs::IntervalSet in
gcs/interval_set.hh, with the forward header and test moving alongside.
The innards keep referring to it unqualified, which now resolves via the
enclosing namespace.

Also make the Problem boundary honest: all_normal_variables() moves from
the "for use by the innards" block to the end-user section (it returns
the variables the user created, and is what the heuristic factories that
take a Problem branch over), and the innards block's documentation now
states the stability policy explicitly.
