--- gcs/innards/interval_set-fwd.hh
@@ -1,10 +1,0 @@
-#ifndef GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_INNARDS_INTERVAL_SET_FWD_HH
-#define GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_INNARDS_INTERVAL_SET_FWD_HH
-
-namespace gcs::innards
-{
-    template <typename Int_>
-    class IntervalSet;
-}
-
-#endif
