+++ gcs/interval_set-fwd.hh
@@ -1,0 +1,10 @@
+#ifndef GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_INTERVAL_SET_FWD_HH
+#define GLASGOW_CONSTRAINT_SOLVER_GUARD_GCS_INTERVAL_SET_FWD_HH
+
+namespace gcs
+{
+    template <typename Int_>
+    class IntervalSet;
+}
+
+#endif
